WebHome: changes after code review
