+++ src/components/Header/index.scss
@@ -3,10 +3,11 @@
 \*------------------------------------*/
 
 .webapp-header {
+  width: 100%;
   height: 50px;
   display: flex;
   background-color: #e7f0f0;
-  padding: 1rem 1rem 2rem 5rem;
+  padding: 1rem 1rem 0rem 5rem;
 
 }
